+++ _app/immutable/entry/start.oQDYhTFZ.js
@@ -1,1 +1,0 @@
-import{r as e,t}from"../chunks/BOsC6FCO.js";export{e as load_css,t as start};